feat(sandbox): 沙箱初始化进度使用后端真实百分比并持久平滑
- progress_percent 由 int 改为 float，子百分比信号可平滑驱动前端
- BootstrapJob 新增 advance_lazy_progress：每次读取时按阶段时间爬升
  （阶段初期快、指数衰减但不为零）并按新增日志行数加速，写回任务本身
- 进度真实、单调且持久：刷新页面不会从 40% 重新跳动
- 前端 SandboxBuildProgress 只显示后端真实值，剩余时间改为从真实
  百分比采样估算，移除前端模拟

+++ backend/app/domain/schemas/sandbox.py
@@ -29,7 +29,7 @@ class SandboxBootstrapStatusView(BaseModel):
     browser_image_digest: str | None = None
     image_source: str | None = None
     phase: str
-    progress_percent: int
+    progress_percent: float
     message: str
     detail: str | None = None
     can_initialize: bool
@@ -58,7 +58,7 @@ class SandboxAgentReadinessView(BaseModel):
 class SandboxBootstrapJobView(BaseModel):
     job_id: str
     phase: str
-    progress_percent: int
+    progress_percent: float
     message: str
     detail: str | None = None
     status: str
